pyproject.toml
[tool.poetry]
name = "insurance-intake-agent"
version = "0.1.0"
description = "AI-powered conversational agent for car insurance registration"
authors = ["Reza Bozorgi"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
gradio = "^5.35.0"
langchain = "^0.3.26"
langchain-openai = "^0.3.27"
langchain-community = "^0.3.26"
motor = "^3.7.1"
dspy-ai = "^2.5.0"
pydantic = "^2.11.7"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
rapidfuzz = "^3.10.0"
pyyaml = "^6.0.2"
openai = "^1.93.0"
transformers = "^4.53.0"
cloudpickle = "^3.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.21.1"
ruff = "^0.12.1"
mypy = "^1.7.1"
types-pyyaml = "^6.0.12.20250516"
types-requests = "^2.32.4.20250611"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
show_error_codes = true
ignore_missing_imports = true
allow_untyped_calls = true
allow_untyped_defs = true
allow_incomplete_defs = true
allow_untyped_decorators = true
no_strict_optional = true

# Per-module configurations for external libraries
[[tool.mypy.overrides]]
module = [
    "dspy.*",
    "transformers.*",
    "gradio.*",
    "motor.*",
    "pymongo.*",
    "langchain.*",
    "langchain_core.*",
    "langchain_openai.*",
    "langchain_community.*",
    "rapidfuzz.*",
    "uvicorn.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]